+++ util/indexer-r/src/indexer/insert.rs
@@ -1,12 +1,11 @@
 #![allow(clippy::needless_borrow)]
 
-use crate::{store::SQLXPool, IndexerRHandle};
+use crate::store::SQLXPool;
 
-use ckb_async_runtime::Handle;
-use ckb_indexer_sync::{CustomFilters, Error, IndexerSync, Pool};
+use ckb_indexer_sync::Error;
 use ckb_types::{
-    core::{BlockNumber, BlockView, TransactionView},
-    packed::{Byte32, OutPointBuilder},
+    core::{BlockView, TransactionView},
+    packed::OutPointBuilder,
     prelude::*,
 };
 use seq_macro::seq;
@@ -14,7 +13,6 @@ use sql_builder::SqlBuilder;
 use sqlx::{Any, Row, Transaction};
 
 use std::collections::HashSet;
-use std::sync::{Arc, RwLock};
 
 // Note that every database has a practical limit on the number of bind parameters you can add to a single query.
 // This varies by database.
@@ -24,136 +22,10 @@ use std::sync::{Arc, RwLock};
 // which should be within the above limits.
 pub(crate) const BATCH_SIZE_THRESHOLD: usize = 1_000;
 
-/// Handle to the indexer-r.
-///
-/// The handle is internally reference-counted and can be freely cloned.
-/// A handle can be obtained using the IndexerRService::handle method.
-///
-/// "Indexer-r" is based on a relational database, and the database tables are as follows:
-///
-/// - block/uncle
-///	    - block_association_proposal
-///	    - block_association_uncle
-/// - tx
-///	    - tx_association_header_dep
-///     - tx_association_cell_dep
-/// - input
-/// - output
-///	    - output_association_script
-/// - script
-///
-/// The detailed table design can be found in the SQL files in the resources folder of this crate.
-
-#[derive(Clone)]
-pub struct IndexerR {
-    async_indexer_r: AsyncIndexerR,
-    async_runtime: Handle,
-}
-
-impl IndexerR {
-    /// Construct new IndexerR instance
-    pub fn new(
-        store: SQLXPool,
-        keep_num: u64,
-        prune_interval: u64,
-        pool: Option<Arc<RwLock<Pool>>>,
-        custom_filters: CustomFilters,
-        async_runtime: Handle,
-    ) -> Self {
-        Self {
-            async_indexer_r: AsyncIndexerR::new(
-                store,
-                keep_num,
-                prune_interval,
-                pool,
-                custom_filters,
-            ),
-            async_runtime,
-        }
-    }
-}
-
-impl IndexerSync for IndexerR {
-    /// Retrieves the tip of the indexer
-    fn tip(&self) -> Result<Option<(BlockNumber, Byte32)>, Error> {
-        let indexer_handle = IndexerRHandle::new(
-            self.async_indexer_r.store.clone(),
-            self.async_indexer_r.pool.clone(),
-            self.async_runtime.clone(),
-        );
-        indexer_handle
-            .get_indexer_tip()
-            .map(|tip| tip.map(|tip| (tip.block_number.value(), tip.block_hash.0.pack())))
-            .map_err(|err| Error::DB(err.to_string()))
-    }
-
-    /// Appends a new block to the indexer
-    fn append(&self, block: &BlockView) -> Result<(), Error> {
-        let future = self.async_indexer_r.append_block(block);
-        self.async_runtime.block_on(future)
-    }
-
-    /// Rollback the indexer to a previous state
-    fn rollback(&self) -> Result<(), Error> {
-        let future = self.async_indexer_r.rollback();
-        self.async_runtime.block_on(future)
-    }
-}
-
-/// Async handle to the indexer-r.
-#[derive(Clone)]
-pub(crate) struct AsyncIndexerR {
-    /// storage
-    store: SQLXPool,
-    /// number of blocks to keep for rollback and forking, for example:
-    /// keep_num: 100, current tip: 321, will prune data where block_number <= 221
-    _keep_num: u64,
-    /// prune interval
-    _prune_interval: u64,
-    /// An optional overlay to index the pending txs in the ckb tx pool
-    /// currently only supports removals of dead cells from the pending txs
-    pool: Option<Arc<RwLock<Pool>>>,
-    /// custom filters
-    _custom_filters: CustomFilters,
-}
-
-impl AsyncIndexerR {
-    /// Construct new AsyncIndexerR instance
-    pub fn new(
-        store: SQLXPool,
-        keep_num: u64,
-        prune_interval: u64,
-        pool: Option<Arc<RwLock<Pool>>>,
-        custom_filters: CustomFilters,
-    ) -> Self {
-        Self {
-            store,
-            _keep_num: keep_num,
-            _prune_interval: prune_interval,
-            pool,
-            _custom_filters: custom_filters,
-        }
-    }
-}
-
-impl AsyncIndexerR {
-    pub(crate) async fn append_block(&self, block: &BlockView) -> Result<(), Error> {
-        let mut tx = self
-            .store
-            .transaction()
-            .await
-            .map_err(|err| Error::DB(err.to_string()))?;
-        insert_block(block, &mut tx).await?;
-        insert_transactions(block, &mut tx).await?;
-        tx.commit().await.map_err(|err| Error::DB(err.to_string()))
-    }
-
-    pub(crate) async fn rollback(&self) -> Result<(), Error> {
-        unimplemented!("rollback")
-    }
-}
-
-async fn insert_block(block_view: &BlockView, tx: &mut Transaction<'_, Any>) -> Result<(), Error> {
+pub(crate) async fn insert_block(
+    block_view: &BlockView,
+    tx: &mut Transaction<'_, Any>,
+) -> Result<(), Error> {
     bulk_insert_block_table(&[block_view.to_owned()], tx).await?;
 
     // bulk_insert_block_association_proposal_table(&[block_view.to_owned()], tx).await?;
@@ -162,7 +34,7 @@ async fn insert_block(block_view: &BlockView, tx: &mut Transaction<'_, Any>) -> 
     Ok(())
 }
 
-async fn insert_transactions(
+pub(crate) async fn insert_transactions(
     block_view: &BlockView,
     tx: &mut Transaction<'_, Any>,
 ) -> Result<(), Error> {
@@ -256,15 +128,15 @@ async fn bulk_insert_block_table(
 }
 
 async fn bulk_insert_block_association_proposal_table(
-    block_views: &[BlockView],
-    tx: &mut Transaction<'_, Any>,
+    _block_views: &[BlockView],
+    _tx: &mut Transaction<'_, Any>,
 ) -> Result<(), Error> {
     unimplemented!("bulk_insert_block_association_proposal_table")
 }
 
 async fn bulk_insert_block_association_uncle_table(
-    block_views: &[BlockView],
-    tx: &mut Transaction<'_, Any>,
+    _block_views: &[BlockView],
+    _tx: &mut Transaction<'_, Any>,
 ) -> Result<(), Error> {
     unimplemented!("")
 }
@@ -546,22 +418,22 @@ async fn bulk_insert_script_table(
 }
 
 async fn bulk_insert_tx_association_header_dep_table(
-    tx_views: &[TransactionView],
-    tx: &mut Transaction<'_, Any>,
+    _tx_views: &[TransactionView],
+    _tx: &mut Transaction<'_, Any>,
 ) -> Result<(), Error> {
     unimplemented!()
 }
 
 async fn bulk_insert_tx_association_cell_dep_table(
-    tx_views: &[TransactionView],
-    tx: &mut Transaction<'_, Any>,
+    _tx_views: &[TransactionView],
+    _tx: &mut Transaction<'_, Any>,
 ) -> Result<(), Error> {
     unimplemented!()
 }
 
 async fn buil_insert_output_association_script(
-    tx_views: &[TransactionView],
-    tx: &mut Transaction<'_, Any>,
+    _tx_views: &[TransactionView],
+    _tx: &mut Transaction<'_, Any>,
 ) -> Result<(), Error> {
     unimplemented!()
 }
