+++ util/indexer-r/src/indexer_handle/mod.rs
@@ -1,4 +1,6 @@
-mod fetch;
+mod async_indexer_handle;
+
+pub use async_indexer_handle::AsyncIndexerRHandle;
 
 use crate::store::SQLXPool;
 
@@ -7,22 +9,6 @@ use ckb_indexer_sync::{Error, Pool};
 use ckb_jsonrpc_types::IndexerTip;
 
 use std::sync::{Arc, RwLock};
-
-/// Async handle to the indexer-r.
-pub struct AsyncIndexerRHandle {
-    store: SQLXPool,
-    _pool: Option<Arc<RwLock<Pool>>>,
-}
-
-impl AsyncIndexerRHandle {
-    /// Construct new AsyncIndexerRHandle instance
-    pub fn new(_store: SQLXPool, _pool: Option<Arc<RwLock<Pool>>>) -> Self {
-        Self {
-            store: _store,
-            _pool,
-        }
-    }
-}
 
 /// Handle to the indexer-r.
 ///
